Last look over CoreL and removed microvias
CoreR has been verified by JLCPCB and is good to go. I forgot to add
teardrops to it so it is now added. CoreL on the other hand had some
more microvias near the JST SH conectors and under the rotary
encoder. These have been removed.

NOTE: Due to the microvia conversion, the previous commit was when the
NTPH holes near the columns have been shifted down by 0.5 mm. This will
be reflected in the case shortly but not yet as of now.

We'll see what JLCPCB says but it looks okay; hopefully I will get
results in a week or so.