5 key variant release!
Well... thought 5 key variant would be a lot of work but it turns out
that it isn't too much. This uses up ALL the pins on the arduino micro
however and unfortunately, a 5 key version for the thumb isn't possible
because one of the keys is used by the encoder.

WARNING: This is NOT compatible with the 4 key version yet. This is
specific to the column PCBs (4-Key and 4Thumb). In the next commit, I
will fix this.