Seismos PCB v1.0rc
First preliminary release of PCB with support for 7 columns and 5 rows.

Folders have now been restructured to organize between cores and mantles.
Cores are the baseboard/motherboard of the keyboard.
Mantles are the connections between the switches and the core.

Of course, additional modules can be made for both cores and mantles.
If you have more pins, larger modules are possible!

Enjoy the first preliminary release!