Added stencils for new and old PCBs
Created stencils for solder pasting PCBs with the core and plates.
I ordered both new and old versions so two stencils were made; however,
the holes for the old variant are misaligned. For this, use the other
aligning holes.

In addition, repaired footprints for switches where the switch footprint
was appearing on the reverse side. This is now fixed.